p-value meta-analysis script
A script to combine p-value across multiple related traits. In a
typical genome-wide association study or a whole-exome sequencing
study, the investigator may want to test for associations with multiple
related traits, e.g., smoking, blood pressure, BMI etc. This script
allows you to combine the p-values across all tested traits under the
null hypothesis that none of the trait is associated with the predictor
(usually a SNV or rare variants in a gene).